# Eeckhoven/bin/Debug/net9.0/swaggerfile.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Eeckhoven",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistrationModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistrationModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistrationModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StringResultObject"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringResultObject"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResultResultObject"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResultResultObject"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResultResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/lesson": {
      "get": {
        "tags": [
          "Lesson"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/lesson/api/lesson/id={id}": {
      "post": {
        "tags": [
          "Lesson"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListResultObject"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListResultObject"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationMessageResultObject"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationMessageResultObject"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationMessageResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/changeLanguage": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResultObject"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResultObject"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/getUser": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserModelResultObject"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModelResultObject"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModelResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/user": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserModelResultObject"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModelResultObject"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModelResultObject"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionResult": {
        "type": "object",
        "additionalProperties": false
      },
      "AddressModel": {
        "required": [
          "city",
          "country",
          "id",
          "state",
          "street",
          "zipCode"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BooleanActionResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ActionResult"
          },
          "value": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "BooleanResultObject": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BooleanActionResult"
            },
            "nullable": true
          },
          "userMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationMessage"
            },
            "nullable": true
          },
          "systemMessages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isError": {
            "type": "boolean"
          },
          "httpErrorCode": {
            "type": "integer",
            "format": "int32"
          },
          "httpErrorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IActionResult": {
        "type": "object",
        "additionalProperties": false
      },
      "IActionResultActionResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ActionResult"
          },
          "value": {
            "$ref": "#/components/schemas/IActionResult"
          }
        },
        "additionalProperties": false
      },
      "IActionResultResultObject": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IActionResultActionResult"
            },
            "nullable": true
          },
          "userMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationMessage"
            },
            "nullable": true
          },
          "systemMessages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isError": {
            "type": "boolean"
          },
          "httpErrorCode": {
            "type": "integer",
            "format": "int32"
          },
          "httpErrorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IdentityUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "JLPT": {
        "enum": [
          "N5",
          "N4",
          "N3",
          "N2",
          "N1",
          "None"
        ],
        "type": "String"
      },
      "Language": {
        "enum": [
          "Japanese",
          "English"
        ],
        "type": "String"
      },
      "LanguageLevel": {
        "enum": [
          "None",
          "Some",
          "Conversational",
          "Business",
          "Fluent",
          "Native"
        ],
        "type": "String"
      },
      "LoginModel": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "remember": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "MessageListActionResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ActionResult"
          },
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationMessage"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MessageListResultObject": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageListActionResult"
            },
            "nullable": true
          },
          "userMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationMessage"
            },
            "nullable": true
          },
          "systemMessages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isError": {
            "type": "boolean"
          },
          "httpErrorCode": {
            "type": "integer",
            "format": "int32"
          },
          "httpErrorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MessageType": {
        "enum": [
          "Warning",
          "Error",
          "Info"
        ],
        "type": "String"
      },
      "RegistrationType": {
        "enum": [
          "User",
          "Company"
        ],
        "type": "String"
      },
      "ResumeModel": {
        "required": [
          "content",
          "id",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/UserModel"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "StringActionResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ActionResult"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StringResultObject": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StringActionResult"
            },
            "nullable": true
          },
          "userMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationMessage"
            },
            "nullable": true
          },
          "systemMessages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isError": {
            "type": "boolean"
          },
          "httpErrorCode": {
            "type": "integer",
            "format": "int32"
          },
          "httpErrorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AddressModel"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "isLoggedIn": {
            "type": "boolean"
          },
          "nationality": {
            "type": "string",
            "nullable": true
          },
          "visaStatus": {
            "$ref": "#/components/schemas/VisaStatus"
          },
          "jlpt": {
            "$ref": "#/components/schemas/JLPT"
          },
          "japaneseLevel": {
            "$ref": "#/components/schemas/LanguageLevel"
          },
          "englishLevel": {
            "$ref": "#/components/schemas/LanguageLevel"
          },
          "profilePicture": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "string",
            "nullable": true
          },
          "identityUser": {
            "$ref": "#/components/schemas/IdentityUser"
          },
          "resume": {
            "$ref": "#/components/schemas/ResumeModel"
          }
        },
        "additionalProperties": false
      },
      "UserModelActionResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ActionResult"
          },
          "value": {
            "$ref": "#/components/schemas/UserModel"
          }
        },
        "additionalProperties": false
      },
      "UserModelResultObject": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserModelActionResult"
            },
            "nullable": true
          },
          "userMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationMessage"
            },
            "nullable": true
          },
          "systemMessages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isError": {
            "type": "boolean"
          },
          "httpErrorCode": {
            "type": "integer",
            "format": "int32"
          },
          "httpErrorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserRegistrationModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AddressModel"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "isLoggedIn": {
            "type": "boolean"
          },
          "nationality": {
            "type": "string",
            "nullable": true
          },
          "visaStatus": {
            "$ref": "#/components/schemas/VisaStatus"
          },
          "jlpt": {
            "$ref": "#/components/schemas/JLPT"
          },
          "japaneseLevel": {
            "$ref": "#/components/schemas/LanguageLevel"
          },
          "englishLevel": {
            "$ref": "#/components/schemas/LanguageLevel"
          },
          "profilePicture": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "string",
            "nullable": true
          },
          "identityUser": {
            "$ref": "#/components/schemas/IdentityUser"
          },
          "resume": {
            "$ref": "#/components/schemas/ResumeModel"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "confirmEmail": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "confirmPassword": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/RegistrationType"
          }
        },
        "additionalProperties": false
      },
      "ValidationMessage": {
        "type": "object",
        "properties": {
          "messageEnglish": {
            "type": "string",
            "nullable": true
          },
          "messageJapanese": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/MessageType"
          }
        },
        "additionalProperties": false
      },
      "ValidationMessageActionResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ActionResult"
          },
          "value": {
            "$ref": "#/components/schemas/ValidationMessage"
          }
        },
        "additionalProperties": false
      },
      "ValidationMessageResultObject": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationMessageActionResult"
            },
            "nullable": true
          },
          "userMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationMessage"
            },
            "nullable": true
          },
          "systemMessages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isError": {
            "type": "boolean"
          },
          "httpErrorCode": {
            "type": "integer",
            "format": "int32"
          },
          "httpErrorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VisaStatus": {
        "enum": [
          "Working",
          "HighlySkilled",
          "IntraCompanyTransfee",
          "Student",
          "WorkingHoliday",
          "Spouse",
          "NoVisa"
        ],
        "type": "String"
      }
    }
  }
}